requirements.txt
asgiref==3.7.2
Django==4.2.7
django-active-link==0.1.8
django-extensions==3.2.3
django-filter==23.3
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3